Add missing features to sigmoid.
The important missing features were: 'raw', 'events, 'cluster-only' and 'host-only.

When sigmoid receives an event, it will send it directly to the graphite service.  For this option to work, the '--env' argument needs to be passed.

For 'raw', 'cluster-only', 'host-only' and 'gauge', we do something slightly different.  In the statsd proxy's protocol, if a line starts with '^(1|2)\s', it knows that it should not multiplex the line.  Since we can't aggregate gauge by cluster, and raw are consider as host-only, when sigmoid is called with this arguments, we append the key with the appropriate mode.

Change-Id: I9c327e37d071e559e662baf330d4271092becbd6